fix: revert sync
